requirements.txt
celery==5.4.0
Django==5.1.1
django-celery-beat==2.7.0
django-celery-results==2.6.0
channels==4.1.0
daphne==4.1.2
django-crispy-forms==2.3
django-environ==0.11.2
django-extensions==3.2.3
django-mysql==4.14.0
djangorestframework==3.15.2
mysqlclient==2.2.4
numpy==2.1.0
pandas==2.2.2
PyMySQL==1.1.1
pytest==8.3.2
pytest-django==4.9.0
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.2.1
sendgrid==6.11.0
sentry-sdk==2.13.0
sqlparse==0.5.1
timezonefinder==6.5.2
uvicorn==0.30.6
gunicorn==23.0.0
whitenoise==6.7.0
django-db-connection-pool==1.2.5
django-redis==6.0.0
boto3==1.40.46
pillow==11.3.0
drf-spectacular==0.28.0